Move running functions into metrics manager
