+++ model_analyzer/record/metrics_manager.py
@@ -12,18 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from model_analyzer.monitor.remote_monitor import RemoteMonitor
-from model_analyzer.constants import LOGGER_NAME
 from .record_aggregator import RecordAggregator
 from .record import RecordType
+from model_analyzer.constants import LOGGER_NAME
 from model_analyzer.device.gpu_device_factory import GPUDeviceFactory
-from model_analyzer.monitor.dcgm.dcgm_monitor import DCGMMonitor
-from model_analyzer.monitor.cpu_monitor import CPUMonitor
-from model_analyzer.perf_analyzer.perf_analyzer import PerfAnalyzer
-from model_analyzer.result.measurement import Measurement
-
 from model_analyzer.model_analyzer_exceptions \
     import TritonModelAnalyzerException
+from model_analyzer.monitor.cpu_monitor import CPUMonitor
+from model_analyzer.monitor.dcgm.dcgm_monitor import DCGMMonitor
+from model_analyzer.monitor.remote_monitor import RemoteMonitor
+from model_analyzer.output.file_writer import FileWriter
+from model_analyzer.perf_analyzer.perf_analyzer import PerfAnalyzer
+from model_analyzer.perf_analyzer.perf_config import PerfAnalyzerConfig
+from model_analyzer.result.measurement import Measurement
 
 from collections import defaultdict
 from prometheus_client.parser import text_string_to_metric_families
@@ -73,6 +74,11 @@ class MetricsManager:
             manages the analyzer state
         """
 
+        # Generate the output model repository path folder.
+        self._output_model_repo_path = config.output_model_repository_path
+
+        self._first_config_variant = None
+
         self._config = config
         self._client = client
         self._server = server
@@ -83,6 +89,10 @@ class MetricsManager:
             self.metrics, self._config.collect_cpu_metrics)
         self._gpus = gpus
         self._init_state()
+
+    def start_new_model(self):
+        """ Indicate that profiling of a new model is starting """
+        self._first_config_variant = None
 
     def _init_state(self):
         """
@@ -149,6 +159,127 @@ class MetricsManager:
             server_gpu_metrics = self._get_gpu_inference_metrics()
             self._result_manager.add_server_data(data=server_gpu_metrics)
         self._destroy_monitors(cpu_only=cpu_only)
+
+    def execute_run_config(self, run_config):
+        """
+        Executes the run config. Returns obtained measurement. Also sends 
+        measurement to the result manager
+        """
+
+        # Create model variant
+        self._create_model_variant(original_name=run_config.model_name(),
+                                   variant_config=run_config.model_config())
+
+        # If this run config was already run, do not run again, just get the measurement
+        measurement = self._get_measurement_if_config_duplicate(run_config)
+        if measurement:
+            return measurement
+
+        # Start server, and load model variant
+        self._server.start(env=run_config.triton_environment())
+        if not self._load_model_variant(
+                variant_config=run_config.model_config()):
+            self._server.stop()
+            return
+
+        # Profile various batch size and concurrency values.
+        # TODO: Need to sort the values for batch size and concurrency
+        # for correct measurment of the GPU memory metrics.
+        perf_output_writer = None if \
+            not self._config.perf_output else FileWriter(self._config.perf_output_path)
+        perf_config = run_config.perf_config()
+
+        logger.info(f"Profiling model {perf_config['model-name']}...")
+        measurement = self.profile_model(run_config=run_config,
+                                         perf_output_writer=perf_output_writer)
+
+        self._server.stop()
+
+        return measurement
+
+    def _create_model_variant(self, original_name, variant_config):
+        """
+        Creates a directory for the model config variant in the output model
+        repository and fills directory with config
+        """
+
+        variant_name = variant_config.get_field('name')
+        if self._config.triton_launch_mode != 'remote':
+            model_repository = self._config.model_repository
+
+            original_model_dir = os.path.join(model_repository, original_name)
+            new_model_dir = os.path.join(self._output_model_repo_path,
+                                         variant_name)
+            try:
+                # Create the directory for the new model
+                os.makedirs(new_model_dir, exist_ok=False)
+                variant_config.write_config_to_file(new_model_dir,
+                                                    original_model_dir,
+                                                    self._first_config_variant)
+                if self._first_config_variant is None:
+                    self._first_config_variant = os.path.join(
+                        self._output_model_repo_path, variant_name)
+            except FileExistsError:
+                pass
+
+    def _load_model_variant(self, variant_config):
+        """
+        Loads a model variant in the client
+        """
+
+        variant_name = variant_config.get_field('name')
+        if self._config.triton_launch_mode != 'c_api':
+            self._client.wait_for_server_ready(self._config.client_max_retries)
+
+            if self._client.load_model(model_name=variant_name) == -1:
+                return False
+
+            if self._client.wait_for_model_ready(
+                    model_name=variant_name,
+                    num_retries=self._config.client_max_retries) == -1:
+                return False
+        return True
+
+    def _get_measurement_if_config_duplicate(self, run_config):
+        """
+        Checks whether this run config has measurements
+        in the state manager's results object
+        """
+
+        model_name = run_config.model_name()
+        model_config_name = run_config.model_config().get_field('name')
+        perf_config_str = run_config.perf_config().representation()
+
+        results = self._state_manager.get_state_variable(
+            'ResultManager.results')
+
+        # check whether perf config string is a key in result dict
+        if model_name not in results:
+            return False
+        if not self._is_config_in_results(
+                run_config.model_config()._model_config, results[model_name]):
+            return False
+        measurements = results[model_name][model_config_name][1]
+
+        # For backward compatibility with keys that still have -u,
+        # we will remove -u from all keys, convert to set and check
+        # perf_config_str is present
+        if perf_config_str in set(
+                map(PerfAnalyzerConfig.remove_url_from_cli_string,
+                    measurements.keys())):
+            return measurements[perf_config_str]
+        else:
+            return None
+
+    def _is_config_in_results(self, config, model_results):
+        """
+        Returns true if `config` exists in the checkpoint `model_results`
+        """
+
+        for result in model_results.values():
+            if config == result[0]._model_config:
+                return True
+        return False
 
     def profile_model(self, run_config, perf_output_writer=None):
         """
